[GPU] Allow precision conversion on get_state() call
