+++ src/plugins/intel_gpu/src/plugin/sync_infer_request.cpp
@@ -62,104 +62,8 @@ inline std::string get_port_name(const ov::Output<const ov::Node>& port, const b
     return name;
 }
 
-template <typename src_t, typename dst_t>
-void convert_any_copy(const src_t* src, dst_t* dst, size_t size) {
-    OPENVINO_ASSERT(src && dst, "[GPU] Src or Dst ptr is null");
-    for (size_t i = 0; i < size; i++)
-        dst[i] = static_cast<dst_t>(src[i]);
-}
-
-void convert_and_copy(const void* src_ptr, ov::element::Type src_et, void* dst_ptr, ov::element::Type dst_et, size_t size) {
-    if (size == 0)
-        return;
-
-    if (src_et == dst_et) {
-        std::memcpy(dst_ptr, src_ptr, size * src_et.size());
-        return;
-    }
-
-    #define CASE(s_et, d_et, s_type, d_type) \
-        if (src_et == s_et && dst_et == d_et) return convert_any_copy(static_cast<const s_type*>(src_ptr), static_cast<d_type*>(dst_ptr), size)
-
-    // For unsupported inputs
-    CASE(ov::element::f64, ov::element::f32, double, float);
-    CASE(ov::element::i16, ov::element::f32, int16_t, float);
-    CASE(ov::element::u16, ov::element::f32, uint16_t, float);
-    CASE(ov::element::u64, ov::element::i32, uint64_t, int32_t);
-    CASE(ov::element::i64, ov::element::i32, int64_t, int32_t);
-    CASE(ov::element::u32, ov::element::i32, uint32_t, int32_t);
-
-    // For unsupported outputs
-    CASE(ov::element::f32, ov::element::f64, float, double);
-    CASE(ov::element::i32, ov::element::i64, int32_t, int64_t);
-    CASE(ov::element::i32, ov::element::u64, int32_t, uint64_t);
-    CASE(ov::element::i32, ov::element::u32, int32_t, uint32_t);
-    CASE(ov::element::f32, ov::element::i16, float, int16_t);
-    CASE(ov::element::f32, ov::element::u16, float, uint16_t);
-
-    // TODO: Need instances below?
-    CASE(ov::element::u32, ov::element::i64, uint32_t, int64_t);
-    CASE(ov::element::u32, ov::element::u64, uint32_t, uint64_t);
-
-    OPENVINO_THROW("[GPU] Unsupported element types combination for copy: ", src_et, " -> ", dst_et);
-}
-
 bool is_convert_required(ov::element::Type src_et, ov::element::Type dst_et) {
     return src_et != dst_et && !(dst_et == ov::element::boolean && src_et == ov::element::u8);
-}
-
-void convert_and_copy(const cldnn::memory::ptr src, ov::ITensor const* dst, const cldnn::stream& stream) {
-    auto src_et = src->get_layout().data_type;
-    auto dst_et = dst->get_element_type();
-
-    size_t size = ov::shape_size(dst->get_shape());
-
-    cldnn::mem_lock<uint8_t> src_lock(src, stream);
-    std::unique_ptr<cldnn::mem_lock<uint8_t>> dst_lock = nullptr;
-
-    const void* src_ptr = src_lock.data();
-    void* dst_ptr = nullptr;
-
-    if (auto remote = dynamic_cast<const ov::intel_gpu::RemoteTensorImpl*>(dst)) {
-        auto mem = remote->get_original_memory();
-        dst_lock.reset(new cldnn::mem_lock<uint8_t>(mem, stream));
-        dst_ptr = dst_lock->data();
-    } else {
-        dst_ptr = dst->data();
-    }
-
-    return convert_and_copy(src_ptr, src_et, dst_ptr, dst_et, size);
-}
-
-void convert_and_copy(const ov::ITensor* src, ov::ITensor const* dst, const cldnn::stream& stream) {
-    auto src_et = src->get_element_type();
-    auto dst_et = dst->get_element_type();
-
-    size_t size = ov::shape_size(dst->get_shape());
-
-    const void* src_ptr = nullptr;
-    void* dst_ptr = nullptr;
-
-    std::unique_ptr<cldnn::mem_lock<uint8_t>> src_lock = nullptr;
-    std::unique_ptr<cldnn::mem_lock<uint8_t>> dst_lock = nullptr;
-
-    if (auto remote = dynamic_cast<const ov::intel_gpu::RemoteTensorImpl*>(src)) {
-        auto mem = remote->get_original_memory();
-        src_lock.reset(new cldnn::mem_lock<uint8_t>(mem, stream));
-        src_ptr = src_lock->data();
-    } else {
-        src_ptr = src->data();
-    }
-
-    if (auto remote = dynamic_cast<const ov::intel_gpu::RemoteTensorImpl*>(dst)) {
-        auto mem = remote->get_original_memory();
-        dst_lock.reset(new cldnn::mem_lock<uint8_t>(mem, stream));
-        dst_ptr = dst_lock->data();
-    } else {
-        dst_ptr = dst->data();
-    }
-
-    return convert_and_copy(src_ptr, src_et, dst_ptr, dst_et, size);
 }
 
 bool same_host_mem(cldnn::memory::cptr memory, const uint8_t* host_ptr) {
