+++ src/plugins/intel_gpu/src/plugin/variable_state.cpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+#include "openvino/core/type/element_type.hpp"
 #include "openvino/runtime/make_tensor.hpp"
 #include "intel_gpu/plugin/remote_context.hpp"
 #include "intel_gpu/plugin/common_utils.hpp"
@@ -19,9 +20,9 @@ namespace intel_gpu {
 VariableState::VariableState(const VariableStateInfo& info, RemoteContextImpl::Ptr context, std::shared_ptr<cldnn::ShapePredictor> shape_predictor)
     : ov::IVariableState {info.m_id}
     , m_layout(info.m_layout)
+    , m_user_specified_type(info.m_user_specified_type)
     , m_context(context)
     , m_shape_predictor(shape_predictor) {
-    m_state = m_context->create_host_tensor(m_layout.data_type, get_tensor_shape(m_layout.get_partial_shape()));
     update_device_buffer();
 }
 
@@ -80,12 +81,16 @@ void VariableState::update_device_buffer() {
     m_memory = m_context->get_engine().reinterpret_buffer(*m_memory, m_layout);
 }
 
-ov::SoPtr<ov::ITensor> VariableState::get_state() const {
-    const bool blocking = true;
-    m_state->set_shape(m_memory->get_layout().get_shape());
-    m_memory->copy_to(m_context->get_engine().get_service_stream(), m_state->data(), blocking);
+ov::element::Type VariableState::get_user_specified_type() const {
+    return m_user_specified_type != ov::element::undefined ? m_user_specified_type : ov::element::Type(m_layout.data_type);
+}
 
-    return m_state;
+ov::SoPtr<ov::ITensor> VariableState::get_state() const {
+    auto tensor = m_context->create_host_tensor(get_user_specified_type(), m_memory->get_layout().get_shape());
+
+    convert_and_copy(m_memory, tensor._ptr.get(), m_context->get_engine().get_service_stream());
+
+    return tensor;
 }
 
 }  // namespace intel_gpu
